==> 01-DataPreprocessing.ipynb <==
In this notebook, we are preparing our dataset for further processing.  Here, we remove null and duplicates.

import numpy as np
import pandas as pd

#loading dataset
urldata = pd.read_csv("Dataset/urldata.csv",encoding = "ISO-8859-1")

urldata.head()

urldata.shape

We want only domain and label columns from the dataset.

domain: URL of a legitimate/phishing website 

Label: 0 indicates legitimate URL

Label: 1 indicates malicious URL

#taking only required columns into a new Dataframe
urldata = urldata.filter(['Domain','Label'])

urldata.head()

urldata.tail()

urldata.shape

#Checking for NULL values
urldata.duplicated().sum()

#removing duplicates
urldata = urldata.drop_duplicates()

urldata.shape

urldata.to_csv("Dataset/preprocessed_urldata.csv",index=False)
