==> 02-Fetch_verified_benign_urls.ipynb <==
In this code we are checking how many benign URLs are reachable out of 47000 benign URLs via whois command.

import numpy as np
import pandas as pd
import whois

import requests
import urllib
import json

urldata = pd.read_csv('Dataset/preprocessed_urldata.csv',encoding = "ISO-8859-1")

len(urldata)

pip install python-whois

#this function performs whois and returns true to those urls which are reachable through whois
def performwhois(url):
    try:
        result = whois.w(url)
        print(result.__dict__)
        return True #success
    except Exception:
        return False #error

benign_sample = urldata['Domain'][1:246]

len(benign_sample)

type(benign_sample)

benign_sample.head()

benign_urls = []
counter = 0
for url in benign_sample:
    if performwhois(url):
        counter = counter + 1
        print(counter)
        benign_urls.append(url)

len(benign_urls)

benign_df = pd.DataFrame()

benign_df['url'] = benign_urls

benign_df['label'] = 0

benign_df.head(245)

benign_df.to_csv("Dataset/whois_verified_benign_urls.csv",index=False)
