==> 03- Sample Code for Attribute Extraction & Preprocessing.ipynb <==
# Basic Libraries to be installed before moving ahead
!pip install pysafebrowsing
!pip install tld
!pip install whois
!pip install geoip2

# Basic Initialisation
import time
import pandas as pd
import numpy as np
pd.set_option('mode.chained_assignment', None) #Switch off warning

#Verifying pathname of dataset before loading
import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename));
        print(os.listdir("../input"))

def loadDataset():
    df = pd.read_csv("/kaggle/input/preprocessing-sample-data/processed_file_3.csv")
    return df

df = loadDataset()
df= df[['url']]
df= df.iloc[0:15,]
df

#Adding columns to the dataframe df
df['ip_add']=""
df['geo_loc']=""
df['tld']=""
df['who_is']=""
df['https']=""
df['label']=""
df = df[['url','ip_add','geo_loc','tld','who_is','https','label']]
df

# Filling the ip_add & geo_loc column of dataframe 
import os
import geoip2.database
import socket
import time

reader = geoip2.database.Reader('/kaggle/input/geolite2countrymmdb/GeoLite2-Country.mmdb')

for x in df.index:
    url = df['url'][x]
    
    try:
        ip_add = socket.gethostbyname(url)
        response = reader.country(ip_add)
        df['geo_loc'][x] = response.country.name
        df['ip_add'][x] = ip_add
        #print(x, "Finished")
        
    except Exception as msg:
        df['geo_loc'][x] = ""
        df['ip_add'][x] = ""
        #print(x," Finished with Error Msg:",msg)

reader.close()

df

#Filling up TLD column
from tld import get_tld

for x in df.index:
        
    try:
        u = df.url[x]
        s = get_tld(str(u), fix_protocol=True)
        df['tld'][x] = s
    except:
        pass
df

#Whois processing
import whois
start_time = time.time()

for x in df.iloc[0:1000].index:
    
    try:    
        domain = whois.query(df['url'][x])
        #print(domain.registrar)
        if len(str(domain.registrar)) >1 :
            df['who_is'][x]= True
        else:
            df['who_is'][x]= False
    except Exception as msg:
        #print(x,", Error: ",msg)
        df['who_is'][x]= False
    print(x,df['who_is'][x])

print("***Total Time taken --- %s seconds ---***" % (time.time() - start_time))
df

# Filling the column https_status
import http.client

start_time = time.time()

for x in df.index:
    https_status= False
    try:
        conn = http.client.HTTPSConnection(df['url'][x])
        conn.request("HEAD", "/")
        res = conn.getresponse()
        if res.status == 200 or res.status==301 or res.status==302:
            https_status= True   
        #print(x,res.status,res.reason,https_status)
    except Exception as msg:
        print(x,"Error: ",msg)
    finally:
        df['https'][x]= https_status
        conn.close

print("***Total Time taken --- %s seconds ---***" % (time.time() - start_time))

df

# Filling the label of training set from Google Safe Browising API
from pysafebrowsing import SafeBrowsing
KEY= "AIzaSyABO6DPGmHpCs8U5ii1Efkp1dUPJHQfGpo"

start_time = time.time()
s = SafeBrowsing(KEY)

for x in df.index:
    
    try:
        url = df['url'][x]
        r = s.lookup_urls([url])
        label=r[url]['malicious']    
        df['label']=label
        #print(x, label)
    except Exception as msg:
        df['label']=""
        #print(x,"Error: ",msg)

print("***Total Time taken --- %s seconds ---***" % (time.time() - start_time))

df

# Saving the file
#df.to_csv("Datasets/processed_1_14Jan20.csv")
