==> 03-Fetch_verified_phishing_urls.ipynb <==
In this code we are checking how many phishing URLs are reachable out of 40000 phishing URLs via whois command.

import numpy as np
import pandas as pd
import whois

urldata = pd.read_csv('Dataset/preprocessed_urldata.csv',encoding = "ISO-8859-1")

len(urldata)

#this function performs whois and returns true to those urls which are reachable through whois
def performwhois(url):
    try:
        result = whois.whois(url)
        return True #success
    except Exception:
        return False #error

phishing_sample = urldata['Domain'][247:3556]

len(phishing_sample)

phishing_sample.head()

#for url in phishing_sample:
    #print(url)

phishing_urls = []
counter = 0
for url in phishing_sample:
    if performwhois(url):
        print(counter)
        counter = counter + 1
        phishing_urls.append(url)

len(phishing_urls)

phishing_urls[0]

phishing_df = pd.DataFrame()

phishing_df["url"] = phishing_urls

phishing_df.head()

phishing_df["label"] = 1

phishing_df.head()

phishing_df.to_csv("Dataset/whois_verified_phishing_urls.csv",index=False)
