==> 04-Merge_verified_urls.ipynb <==
In this notebook, we have taken about 5000 urls of each whois verified benign urls and whois verified phishing urls to form the final dataset.

import pandas as pd

phishing_urls_whois_verified = pd.read_csv("Dataset/whois_verified_phishing_urls.csv")

benign_urls_whois_verified = pd.read_csv("Dataset/whois_verified_benign_urls.csv")

phishing_urls_whois_verified.shape

benign_urls_whois_verified.shape

benign_urls_whois_verified['url']= benign_urls_whois_verified['url'].apply(lambda x: "{}{}".format('https://', x))

phishing_urls_whois_verified.to_csv("Dataset/DTphish.csv",index=False)

phishing_urls_whois_verified['url']= phishing_urls_whois_verified['url'].apply(lambda x: "{}{}".format('http://', x))

phishing_urls_whois_verified.head()

phishing_urls_whois_verified.to_csv("Dataset/DTphish.csv",index=False)

benign_urls_whois_verified.head()

phishing_urls_whois_verified_reduced = phishing_urls_whois_verified[:3000]

phishing_urls_whois_verified_reduced.shape

benign_urls_whois_verified_reduced = benign_urls_whois_verified[:245]

benign_urls_whois_verified_reduced.shape

combined = pd.DataFrame()

combined = pd.concat([benign_urls_whois_verified_reduced,phishing_urls_whois_verified_reduced])

combined.shape

combined.to_csv("Dataset/PhishingData.csv", index=False)

# Load Datasets
def loadDataset(file_name):
    df = pd.read_csv(file_name)
    return df

df = loadDataset("Dataset/PhishingData.csv")

df.info

df.head()

df.tail()

df.to_csv('Dataset/data.csv',index=False)
