==> 10-I-CICBELLDataPreprocessing.ipynb <==
In this notebook, we are preparing our dataset for further processing.  Here, we remove null and duplicates.

import numpy as np
import pandas as pd

# Load Datasets
def loadDataset(file_name):
    df = pd.read_csv(file_name, encoding = "ISO-8859-1", on_bad_lines='skip')
    return df

df = loadDataset("CICBELLDNS2021datasets/CSV_phishing.csv") 

df.info()

df.shape

df.columns

We want only domain and label columns from the dataset.

## domain of a legitimate/phishing website 
 
## Label: 0 indicates legitimate URL

## Label: 1 indicates malicious URL

#taking only required columns into a new Dataframe
#urldata = urldata.filter(['Domain','Label'])

df

#Adding columns label = 1 for phishing domains to the dataframe df
df['label']="1"
df
df = df[['Country', 'ASN', 'TTL', 'IP', 'Domain', 'State', 'Registrant_Name','Country.1', 'Creation_Date_Time', 'hex_32', 'hex_8', 'Domain_Name',
       'Alexa_Rank', 'subdomain', 'Organization', 'len', 'longest_word',
       'oc_32', 'shortened', '1gram', 'obfuscate_at_sign', 'entropy',
       'Domain_Age', 'tld', 'dec_8', 'dec_32', 'Emails', 'numeric_percentage',
       'puny_coded', 'typos', 'oc_8', '3gram', 'char_distribution', '2gram',
       'Registrar', 'sld', 'Name_Server_Count', 'Page_Rank','label']]

df

df.shape

urldata= loadDataset("CICBELLDNS2021datasets/CSV_benign.csv") 

urldata.info()

urldata

urldata.shape

#removing duplicates
urldata = urldata.drop_duplicates()

#urldata.shape

urldata.to_csv("CICBELLDNS2021datasets/processed_CSV_benign.csv",index=False)

bd = loadDataset("CICBELLDNS2021datasets/processed_CSV_benign.csv") 

bd.shape

bd.info()

bd

#Adding columns label = 0 for benign domains to the dataframe 
bd['label']="0"
bd
bd = bd[['Country', 'ASN', 'TTL', 'IP', 'Domain', 'State', 'Registrant_Name','Country.1', 'Creation_Date_Time', 'hex_32', 'hex_8', 'Domain_Name',
       'Alexa_Rank', 'subdomain', 'Organization', 'len', 'longest_word',
       'oc_32', 'shortened', '1gram', 'obfuscate_at_sign', 'entropy',
       'Domain_Age', 'tld', 'dec_8', 'dec_32', 'Emails', 'numeric_percentage',
       'puny_coded', 'typos', 'oc_8', '3gram', 'char_distribution', '2gram',
       'Registrar', 'sld', 'Name_Server_Count', 'Page_Rank','label']]

bd

bd.shape

#495663+5001

# using merge function by setting how='inner'
df_final = pd.concat([df,bd])

df_final.shape

df_final.info()

df_final.to_csv("Train_Dataset/Training_data_INPDS.csv",index=False)
